Made code compile with all bonuses working

+++ CODE/utility/definition.h
@@ -1,7 +1,7 @@
 typedef struct Instance{
   char id[81];
   int max; //Used only in histo commands
-  int volume; //Corresponds to "src" in histo commands
+  double volume; //Corresponds to "src" in histo commands
   double leaks; //Corresponds to "real" in histo commands
   struct Instance** downstream;
   int downstreamCount;
